refactor: 패키지 및 클래스 이름 변경
